requirements.txt
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
Flask-Mail==0.10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
paypalrestsdk==1.13.3
pycparser==2.22
PyMySQL==1.1.1
pyOpenSSL==25.0.0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
stripe==11.4.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
